Remove Neutron external network interface config
In some legacy deployments without firewalls we used the controllers as
gateways for the Neutron external and internal pNodes networks. All new
deployments use external network infrastructure to provide the gateway
for these networks. Remove code to automatically generate an
interfaces(5) file for the brq* interface for the external Neutron
network. Existing deployments will be unaffected since this file is
already written, and in most cases defined host_vars network_interfaces.
This will prevent erroneously assigning a duplicate IP address to the
controllers as the external default gateway.